refactor(croquis-cf): rename rule implementation module
