+++ crates/vize_croquis_cf/src/rules/boundary.rs
