+++ crates/vize_croquis_cf/src/rules/component_resolution.rs
@@ -7,6 +7,10 @@ use crate::graph::DependencyGraph;
 use crate::registry::{FileId, ModuleRegistry};
 use vize_carton::{CompactString, FxHashSet, String, cstr};
 
+mod paths;
+
+use paths::resolve_import;
+
 /// Information about a component resolution issue.
 #[derive(Debug, Clone)]
 pub struct ComponentResolutionIssue {
@@ -258,235 +262,5 @@ fn to_pascal_case(value: &str) -> String {
         .collect()
 }
 
-/// Try to resolve an import specifier to a file in the registry.
-#[allow(clippy::disallowed_macros)]
-fn resolve_import(
-    specifier: &str,
-    registry: &ModuleRegistry,
-    from_dir: Option<&std::path::Path>,
-) -> bool {
-    // Handle @/ alias (common Vue project alias)
-    if let Some(relative) = specifier.strip_prefix("@/") {
-        // Check with common extensions
-        for ext in &["", ".vue", ".ts", ".tsx", ".js", ".jsx"] {
-            let path = format!("src/{}{}", relative, ext);
-            if registry.get_by_path(&path).is_some() {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    // Handle relative imports.
-    //
-    // Resolve the specifier against the importing file's directory and
-    // normalize `.`/`..` segments to a canonical path. Matching only the
-    // canonical path ensures `./Button.vue` resolves to the sibling file and
-    // never to a different directory's same-named file (e.g. `admin/Button.vue`).
-    if specifier.starts_with('.') {
-        let candidates = import_candidates(specifier, from_dir);
-        return registry.iter().any(|entry| {
-            let entry_path = normalize_logical_path(entry.path.clone());
-            candidates.contains(&entry_path)
-        });
-    }
-
-    // For absolute or other paths, check directly
-    registry.get_by_path(specifier).is_some()
-}
-
-/// Build the set of canonical candidate paths a relative import specifier may
-/// resolve to, trying the common module extensions and `index` files.
-fn import_candidates(
-    specifier: &str,
-    from_dir: Option<&std::path::Path>,
-) -> Vec<std::path::PathBuf> {
-    use std::path::PathBuf;
-
-    let base = from_dir
-        .filter(|dir| !dir.as_os_str().is_empty())
-        .map_or_else(|| PathBuf::from(specifier), |dir| dir.join(specifier));
-
-    let mut candidates = Vec::new();
-    let has_extension = base.extension().is_some();
-    candidates.push(normalize_logical_path(base.clone()));
-
-    if !has_extension {
-        for suffix in [
-            ".vue",
-            ".ts",
-            ".tsx",
-            ".js",
-            ".jsx",
-            "/index.vue",
-            "/index.ts",
-            "/index.tsx",
-            "/index.js",
-            "/index.jsx",
-        ] {
-            candidates.push(normalize_logical_path(path_with_suffix(&base, suffix)));
-        }
-    }
-
-    candidates
-}
-
-fn path_with_suffix(base: &std::path::Path, suffix: &str) -> std::path::PathBuf {
-    if let Some(index_file) = suffix.strip_prefix('/') {
-        base.join(index_file)
-    } else {
-        let mut value = base.as_os_str().to_os_string();
-        value.push(suffix);
-        std::path::PathBuf::from(value)
-    }
-}
-
-/// Normalize a path by collapsing `.`/`..` segments without touching the
-/// filesystem, yielding a canonical logical path for comparison.
-fn normalize_logical_path(path: std::path::PathBuf) -> std::path::PathBuf {
-    use std::path::{Component, Path, PathBuf};
-
-    let mut normalized = PathBuf::new();
-
-    for component in path.components() {
-        match component {
-            Component::CurDir => {}
-            Component::ParentDir => {
-                normalized.pop();
-            }
-            Component::Normal(part) => normalized.push(part),
-            Component::RootDir => normalized.push(Path::new("/")),
-            Component::Prefix(prefix) => normalized.push(prefix.as_os_str()),
-        }
-    }
-
-    normalized
-}
-
 #[cfg(test)]
-mod tests {
-    use super::{
-        ComponentResolutionIssueKind, analyze_component_resolution, is_builtin_component,
-        is_custom_element_tag, resolve_import,
-    };
-    use crate::CrossFileDiagnosticKind;
-    use crate::graph::DependencyGraph;
-    use crate::registry::ModuleRegistry;
-    use vize_carton::{CompactString, smallvec};
-    use vize_croquis::analysis::ComponentUsage;
-    use vize_croquis::{Croquis, ScopeId};
-
-    #[test]
-    fn test_is_builtin_component() {
-        assert!(is_builtin_component("Transition"));
-        assert!(is_builtin_component("transition"));
-        assert!(is_builtin_component("transition-group"));
-        assert!(is_builtin_component("KeepAlive"));
-        assert!(is_builtin_component("keep-alive"));
-        assert!(is_builtin_component("RouterView"));
-        assert!(is_builtin_component("router-view"));
-        assert!(is_builtin_component("NuxtPage"));
-        assert!(is_builtin_component("nuxt-page"));
-        assert!(is_builtin_component("nuxt-link"));
-        assert!(is_builtin_component("client-only"));
-        assert!(is_builtin_component("slot"));
-        assert!(!is_builtin_component("MyComponent"));
-        assert!(!is_builtin_component("UserCard"));
-    }
-
-    #[test]
-    fn test_is_custom_element_tag() {
-        assert!(is_custom_element_tag("my-widget"));
-        assert!(is_custom_element_tag("ion-button"));
-        assert!(is_custom_element_tag("sl-icon2"));
-        assert!(is_custom_element_tag("my_widget-button"));
-        assert!(!is_custom_element_tag("MyWidget"));
-        assert!(!is_custom_element_tag("myWidget"));
-        assert!(!is_custom_element_tag("ChildWidget"));
-        assert!(!is_custom_element_tag("font-face"));
-        assert!(!is_custom_element_tag("div"));
-    }
-
-    #[test]
-    fn unregistered_component_uses_template_usage_offset() {
-        let mut registry = ModuleRegistry::new();
-        let graph = DependencyGraph::new();
-        let mut analysis = Croquis::new();
-
-        analysis
-            .used_components
-            .insert(CompactString::new("UnknownThing"));
-        analysis.component_usages.push(ComponentUsage {
-            name: CompactString::new("UnknownThing"),
-            start: 12,
-            end: 27,
-            props: smallvec![],
-            events: smallvec![],
-            slots: smallvec![],
-            has_spread_attrs: false,
-            scope_id: ScopeId::ROOT,
-            vif_guard: None,
-        });
-
-        let (file_id, _) = registry.register("Parent.vue", "", analysis);
-        let (issues, diagnostics) = analyze_component_resolution(&registry, &graph);
-
-        assert_eq!(issues.len(), 1);
-        assert_eq!(issues[0].file_id, file_id);
-        assert_eq!(
-            issues[0].kind,
-            ComponentResolutionIssueKind::UnregisteredComponent
-        );
-        assert_eq!(issues[0].offset, 12);
-
-        assert_eq!(diagnostics.len(), 1);
-        assert_eq!(diagnostics[0].primary_file, file_id);
-        assert_eq!(diagnostics[0].primary_offset, 12);
-        assert_eq!(
-            diagnostics[0].kind,
-            CrossFileDiagnosticKind::UnregisteredComponent {
-                component_name: CompactString::new("UnknownThing"),
-                template_offset: 12,
-            }
-        );
-    }
-
-    #[test]
-    fn relative_import_resolves_to_sibling_not_same_named_file_elsewhere() {
-        let mut registry = ModuleRegistry::new();
-
-        // Two components named `Button.vue` in different directories.
-        registry.register("pages/Button.vue", "", Croquis::new());
-        registry.register("admin/Button.vue", "", Croquis::new());
-
-        // `./Button.vue` imported from `pages/Home.vue` must resolve to the
-        // sibling `pages/Button.vue`.
-        let from_dir = std::path::Path::new("pages/Home.vue").parent();
-        assert!(resolve_import("./Button.vue", &registry, from_dir));
-    }
-
-    #[test]
-    fn relative_import_does_not_cross_directories() {
-        let mut registry = ModuleRegistry::new();
-
-        // Only the `admin/` variant exists; the sibling does not.
-        registry.register("admin/Button.vue", "", Croquis::new());
-
-        // `./Button.vue` imported from `pages/Home.vue` must NOT resolve to
-        // `admin/Button.vue` via a suffix match.
-        let from_dir = std::path::Path::new("pages/Home.vue").parent();
-        assert!(!resolve_import("./Button.vue", &registry, from_dir));
-    }
-
-    #[test]
-    fn relative_import_resolves_parent_directory() {
-        let mut registry = ModuleRegistry::new();
-
-        registry.register("components/Button.vue", "", Croquis::new());
-
-        // `../Button.vue` from `components/forms/Field.vue` resolves to
-        // `components/Button.vue`.
-        let from_dir = std::path::Path::new("components/forms/Field.vue").parent();
-        assert!(resolve_import("../Button.vue", &registry, from_dir));
-    }
-}
+mod tests;
