+++ crates/vize_croquis_cf/src/rules/cross_file_reactivity/collect.rs
