+++ crates/vize_croquis_cf/src/rules/emit.rs
