+++ crates/vize_croquis_cf/src/rules/provide_inject/index.rs
