+++ crates/vize_croquis_cf/src/rules/reactivity/component.rs
