+++ crates/vize_croquis_cf/src/rules/reactivity/diagnostics.rs
