+++ crates/vize_croquis_cf/src/rules/reactivity/losses.rs
