+++ crates/vize_croquis_cf/src/rules/reactivity/mod.rs
