+++ crates/vize_croquis_cf/src/rules/reactivity/types.rs
