+++ crates/vize_croquis_cf/src/rules/setup_context.rs
